Editing and rebuilding od test structures, especially integration-test. Flyway migration configuration

+++ KlubFitness/src/main/resources/db/migration/V1__init.sql
@@ -1,6 +1,6 @@
 -- users
 CREATE TABLE users (
-                       id SERIAL PRIMARY KEY,
+                       id BIGSERIAL PRIMARY KEY,
                        username VARCHAR(255) UNIQUE NOT NULL,
                        password VARCHAR(255) NOT NULL,
                        role VARCHAR(50) NOT NULL
@@ -8,25 +8,28 @@ CREATE TABLE users (
 
 -- trainers
 CREATE TABLE trainers (
-                          id SERIAL PRIMARY KEY,
+                          id BIGSERIAL PRIMARY KEY,
                           name VARCHAR(255) NOT NULL,
                           specialization VARCHAR(255)
 );
 
 -- training_sessions
 CREATE TABLE training_sessions (
-                                   id SERIAL PRIMARY KEY,
+                                   id BIGSERIAL PRIMARY KEY,
                                    title VARCHAR(255) NOT NULL,
                                    description TEXT,
                                    start_time TIMESTAMP NOT NULL,
                                    end_time TIMESTAMP NOT NULL,
-                                   trainer_id INT NOT NULL REFERENCES trainers(id)
+                                   trainer_id BIGINT NOT NULL
+                                       REFERENCES trainers(id)
 );
 
 -- reservations
 CREATE TABLE reservations (
-                              id SERIAL PRIMARY KEY,
+                              id BIGSERIAL PRIMARY KEY,
                               reservation_time TIMESTAMP NOT NULL,
-                              user_id INT NOT NULL REFERENCES users(id),
-                              session_id INT NOT NULL REFERENCES training_sessions(id)
+                              user_id BIGINT NOT NULL
+                                  REFERENCES users(id),
+                              session_id BIGINT NOT NULL
+                                  REFERENCES training_sessions(id)
 );
